Extracted Outbound Init
The Outbound init call was in the MainActivity but it should be in the
Application class. I created an application class, inserted the call to
init, and added the class to the Android Manifest.
