+++ sampleapp/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.zendesk.connect.sampleapp">
 
     <application
+        android:name=".SampleApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
